fix: correct import paths in writer module to use restructured packages
- handlers.py: Update 4 imports from deleted packages
  - egregora.augmentation.profiler → egregora.agents.tools.profiler
  - egregora.knowledge.annotations → egregora.agents.tools.annotations
  - egregora.knowledge.rag → egregora.agents.tools.rag
  - egregora.orchestration.write_post → egregora.utils.write_post

- formatting.py: Update annotation imports
  - egregora.knowledge.annotations → egregora.agents.tools.annotations

These were the last remaining references to the old package structure
that were missed during the initial restructure.

Fixes: https://github.com/franklinbaldo/egregora/pull/587#discussion_r2495012946
Fixes: https://github.com/franklinbaldo/egregora/pull/587#discussion_r2495012957
