+++ src/egregora/agents/writer/handlers.py
@@ -8,12 +8,12 @@ from typing import Any
 from google.genai import types as genai_types
 
 from egregora.agents.banner import generate_banner_for_post
+from egregora.agents.tools.annotations import AnnotationStore
+from egregora.agents.tools.profiler import read_profile, write_profile
+from egregora.agents.tools.rag import VectorStore, query_media
 from egregora.agents.writer.formatting import _stringify_value
-from egregora.augmentation.profiler import read_profile, write_profile
-from egregora.knowledge.annotations import AnnotationStore
-from egregora.knowledge.rag import VectorStore, query_media
-from egregora.orchestration.write_post import write_post
 from egregora.utils import GeminiBatchClient
+from egregora.utils.write_post import write_post
 
 logger = logging.getLogger(__name__)
